+++ EssentialFeed/EssentialFeed/Feed API/FeedItemsMapper.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-internal final class FeedItemsMapper {
+public final class FeedItemsMapper {
     private struct Root: Decodable {
         private let items: [RemoteFeedItem]
         
@@ -16,7 +16,7 @@ internal final class FeedItemsMapper {
         }
     }
 
-    internal static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
+    public static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
         guard response.isOK, let root = try? JSONDecoder().decode(Root.self, from: data) else {
             throw RemoteFeedLoader.Error.invalidData
         }
